Gui Control of manager and dataserver + save
Finalized the GUI control of the manager and the data server.
Connections can now be created and cancelled using the manager GUI.

Also changed the format of the data sent by the acquire loop and how the
save function handles that new format. I also noticed a bit of a bug;
initial values of p in the acquire loop were ints, and were then set to
floats when scanning. This made the HFS store fail.