Added setpoint
Updated (a few) of the scripts to PEP8. Added labels to the Manager GUI
to differentiate scanning and setpoint controls. Sending a setpoint
should work, have not tested it with an active Artist (but the correct
message should be sent).
Setpoint works by setting a value, and measuring for 50 ms. I don't know
if this distinction matters (could be set to 0 seconds, for instance).
If wanted, an additional time box can be added to specify the
measurement time.